requirements.txt
discord.py
python-dotenv
emoji
# Isthereanydeal
requests
beautifulsoup4